+++ src/system.ts
@@ -14,3 +14,4 @@ export * from './system/iterable';
 export * from './system/object';
 export * from './system/searchTree';
 export * from './system/string';
+export * from './system/version';